refactored image blocks save image functions
